refactor(scripts): make maintenance scripts MySQL-compatible
Updates utility and data scripts to remove returning/cast assumptions, use explicit IDs where needed, and keep helper flows working with the MySQL schema.

+++ scripts/recalculate-post-counts.js
@@ -13,7 +13,7 @@ async function recalculatePostCounts() {
   for (const category of allCategories) {
     // Count published posts in this category
     const [{ count }] = await db
-      .select({ count: sql`count(*)::integer` })
+      .select({ count: sql`count(*)` })
       .from(posts)
       .where(eq(posts.categoryId, category.id))
       .where(eq(posts.status, 'PUBLISHED'));
